decomp: stage recovered C into 55 pending units' NON_MATCHING branches
Each file keeps its retail oracle under #ifndef NON_MATCHING, untouched - this
commit only replaces the #else branch, which the matching build never compiles.
Zero INCLUDE_ASM lines were removed, which is the check that the oracle survived
intact. The full baseline still passes: reconstructed boot ELF matches retail,
byte for byte.

This is where the campaign's recovered work becomes visible in the public tree.
Until now the recoveries lived in the private candidate bank and the public
progress report could not see them, so 168 units sat there unnoticed.

55 of them, and the mechanisms that produced the biggest moves are recorded where
a reader can find them - in analysis/functions/<unit>.md in the tools repo,
which is the campaign's documentation, rather than in a comment here:

- cpr8 reached C_EXACT (promoted separately) on four pure compiler policies:
  volatile GS/DMA register stores, while-form spin loops over goto self-loops,
  a while-form outer loop that spills arg0, and a for-form inner loop that flips
  s0/s1.
- sceFsInit 96.26, from a retry loop the oracle's block placement revealed
  (bind -> queue empty -> countdown -> bind) that freed the sema addresses.
- sceOpen 91.07, where a variadic frame is the only thing that produces the six
  sd a2..t3,0x120..0x148 register saves and the 0x150 frame on this line.
- _getPtsDtsFlags 97.78, almost entirely from struct widths: 0x78/0x88/0xF0 and
  arg1+0x18/0x20 are s64.
- sce_lseek 94.99, and the bank that scored 88.51 carried three forbidden
  register __asm__ pins - the pin-free source beats it outright.
- sceCdRead 90.25, where dropping the invented return-value variable was worth
  +9.25, because the oracle has no phi.
- sce_tty_write 93.61, sce_sif_init_cmd 90.93, sceWrite 85.04, disprefimagefield
  96.18 (its 96.31 bank was entirely pin-based), _updateRefImage 93.84,
  _getAllRefs 89.18, _PES_packet 92.41, sceCdInit 70.76, printfloat 91.89.

Three real m2c semantic errors surfaced and are corrected here: _getAllRefs
tested arg4 == 2 where retail tests arg4 == 3; _updateRefImage had the
merged-pointer select swapped in block_24; and cpr8's two call sites take no
arguments at all, because a0/a1 still hold the addresses the preceding store
block built.

+++ src/assembly/textbin/fun_002157d0.c
@@ -12,7 +12,7 @@ extern u8 D_00137B80[];
 extern struct M2c_D_001516D0 D_001516D0;
 extern u8 D_002169C0[];
 extern s32 func_0012EC08();
-void FUN_002157d0(s16 arg0, s16 arg1, s16 arg2) {
+void FUN_002157d0(s32 arg0, s32 arg1, s32 arg2) {
     s32 temp_4_12;
 
     temp_4_12 = *(s32 *)((u8 *)(D_00137B80 + ((arg0 - 0x4E20) * 8)) + 0x2988);
